+++ src/components/FolderList/FolderList.jsx
@@ -2,20 +2,15 @@ import { useState } from "react";
 import * as S from "./FolderList.style";
 import addIcon from "images/add.svg";
 import addIconWhite from "images/add_white.svg";
-import deleteIcon from "images/delete.svg";
-import nameChangeIcon from "images/name-change.svg";
-import shareIcon from "images/share.svg";
 import Modal from "components/Modal";
-import { ModalAddFolder, ModalDelete, ModalEdit, ModalShare } from "components/Modal/Modal";
 import useRequest from "hooks/useRequest";
+import ModalAddFolder from "components/Modal/ModalAddFolder";
+import CurrentFolderInfo from "components/CurrentFolderInfo";
 
 function FolderList({ getFolderId }) {
   const [selectedId, setSelectedId] = useState(null);
   const [selectedName, setSelectedName] = useState("전체");
-  const [modalIsOpen1, setModalIsOpen1] = useState(false);
-  const [modalIsOpen2, setModalIsOpen2] = useState(false);
-  const [modalIsOpen3, setModalIsOpen3] = useState(false);
-  const [modalIsOpen4, setModalIsOpen4] = useState(false);
+  const [modalIsOpen, setModalIsOpen] = useState(false);
 
   const { data } = useRequest({ url: "/users/1/folders" });
   const folders = data?.data;
@@ -46,50 +41,19 @@ function FolderList({ getFolderId }) {
               </S.Folder>
             ))}
           </S.FolderContainer>
-          <S.AddFolderButton onClick={() => setModalIsOpen4(true)}>
+          <S.AddFolderButton onClick={() => setModalIsOpen(true)}>
             <span>폴더 추가</span>
             <img className="notMobile" src={addIcon} alt="추가 아이콘" />
             <img className="onMobile" src={addIconWhite} alt="추가 아이콘" />
           </S.AddFolderButton>
-          {modalIsOpen4 && (
-            <Modal close={() => setModalIsOpen4(false)}>
+          {modalIsOpen && (
+            <Modal close={() => setModalIsOpen(false)}>
               <ModalAddFolder />
             </Modal>
           )}
         </S.FolderListContainer>
       )}
-      <S.CurrentFolderInfo>
-        <span>{selectedName}</span>
-        <S.OptionContainer selected={!selectedId}>
-          <button onClick={() => setModalIsOpen1(true)}>
-            <img src={shareIcon} alt="공유 아이콘" />
-            <span>공유</span>
-          </button>
-          {modalIsOpen1 && (
-            <Modal close={() => setModalIsOpen1(false)}>
-              <ModalShare folderName={selectedName} folderId={selectedId}/>
-            </Modal>
-          )}
-          <button onClick={() => setModalIsOpen2(true)}>
-            <img src={nameChangeIcon} alt="이름 변경 아이콘" />
-            <span>이름 변경</span>
-          </button>
-          {modalIsOpen2 && (
-            <Modal close={() => setModalIsOpen2(false)}>
-              <ModalEdit folderName={selectedName} />
-            </Modal>
-          )}
-          <button onClick={() => setModalIsOpen3(true)}>
-            <img src={deleteIcon} alt="삭제 아이콘" />
-            <span>삭제</span>
-          </button>
-          {modalIsOpen3 && (
-            <Modal close={() => setModalIsOpen3(false)}>
-              <ModalDelete folderName={selectedName} />
-            </Modal>
-          )}
-        </S.OptionContainer>
-      </S.CurrentFolderInfo>
+      <CurrentFolderInfo selectedName={selectedName} selectedId={selectedId} />
     </>
   );
 }
