+++ src/components/SampleCardList/SampleCard/SampleCard.jsx
@@ -2,7 +2,7 @@ import * as S from "./SampleCard.style";
 import noImage from "images/no-image.svg";
 import star from "images/star.svg";
 import kebab from "images/kebab.svg";
-import { formatDate, formatTimeDiff } from "utils";
+import { formatDate, formatTimeDiff } from "utils/format";
 
 function SampleCard({ item }) {
   const { imageSource, createdAt, title, description, url } = item;
